+++ packages/ui/components/core/Badge.stories.tsx
@@ -1,6 +1,7 @@
-import { Badge } from '@mantine/core'
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 import React from 'react'
+
+import { Badge } from './Badge'
 
 export default {
 	title: 'Design System/Tags and Badges',
