--- packages/ui/theme/functions/index.ts
@@ -1,3 +1,0 @@
-// codegen:start {preset: barrel, include: ./*.ts*}
-export * from './badgeVariants'
-// codegen:end
